Changed probability to double

+++ src/model/Movement.java
@@ -1,18 +1,18 @@
 package model;
 
 public class Movement {
-	int probability;
+	double probability;
 	int[] pos;
 	
 	public Movement(double probability, int row, int col, int head){
-		this.probability = (int) (probability*100);
+		this.probability =  (probability*100);
 		pos = new int[3];
 		pos[0] = row;
 		pos[1] = col;
 		pos[2] = head;
 	}
 	
-	public int relativeProb(){
+	public double relativeProb(){
 		return probability;
 	}
