+++ src/model/Reading.java
@@ -1,17 +1,17 @@
 package model;
 
 public class Reading {
-	int probability;
+	double probability;
 	int[] pos;
 	
 	public Reading(double probability, int row, int col){
-		this.probability = (int) (probability*100);
+		this.probability = (probability*100);
 		pos = new int[2];
 		pos[0] = row;
 		pos[1] = col;
 	}
 	
-	public int relativeProb(){
+	public double relativeProb(){
 		return probability;
 	}
